Combining Status and StatusOr into a single file. This will make it easier to rearrange the dependencies.
